Add files via upload
Included are files for logging in/out a user (index is initial login, login/logout perform actions, home is the redirect after logged in (where book list would appear), and the database files are for the DB connection. There are still a few bugs, and I'll need to clean up security issues, but it's at least a working model. The bookshare.sql file contains the user table used in the login process, with one user entered (me).