Per Professor Dietz's request, added input validation to signup page:
(1) no spaces in the username, 
(2) password should contain a capital letter and a number, 
(3) the email address should have an @ sign.  

The email and username are also checked against the database for already existing entries.

If a problem is found, the following flags (boolean) are set using session, and the signup view is called again.

err_email
err_username
err_password
err_duplicate

Right now, this displays a vague error message using PHP, but front end should feel free to revise as needed.

The signup page has also been moved to a separate html document with a corresponding php page to carry out the action. All headers in index.php have been moved to separate pages that are included via PHP to allow different headers based on session data. The mobile versions have been incorporated, and I edited which links appear in the headers (for example, My Profile should not appear if no user is logged in).